+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 - Required binary data can now be pushed to workers.
 - `ConversationTransformers` can now push `MessageRead` and `ConversationRead` events.
+- Fix unhandled exception on missing artifact (now a warning)
+- Allow FileConcatenate to create new directories and overwrite existing files.
 
 ---
 *0.1.0-alpha.4:*