+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - `ConversationTransformers` can now push `MessageRead` and `ConversationRead` events.
 - Fix unhandled exception on missing artifact (now a warning)
 - Allow FileConcatenate to create new directories and overwrite existing files.
+- ExactReplace can now use external csv data (using the new `PackedDataCache`)
 
 ---
 *0.1.0-alpha.4:*